Add command line option --route-reuse to override the default setting of 10 to experiment with better route reuse settings
Adjust the TradeCalculation distance weights to give a better result.

Adjust some of the sort vs sorted uses for cleaner code. 

Increase the number of worlds left in the stars list used for
determining routes. It was too restrictive and was removing too may
links. May need more exact system for sorting to get better results.