Supply absolute wiki template root path
Previously, the template path was a relative one - taken from the
current working directory.  This is fine when running PyRoute from
the command line, but not necessarily so under an IDE.

For example, PyCharm on Fedora 33 changes the working dir to
something hanging off its own install dir, _instead_ of the project
directory.  As a result, when running or profiling in PyCharm, the
FileSystemLoader is sent haring off on a wild goose chase, it
can't find the templates, and the run blows up.

The simple approach, and the one I've taken, is to remove that
latent dependency and explicitly supply the absolute path template.