Pre-empt doomed-to-failure pathfinding attempts
Due to the astrography implied in the sector maps being read in,
it seems fairly rare for the galaxy.stars graph to be connected -
ie, each and every star reachable via pathfinding from eacn and
every other.

Pathfinding attempts between such disconnected components are
both doomed to failure (as no such underlying path exists),
and disproportionately more expensive than average.

As we're pulling in networkX for the pathfinding _anyway_,
why not use another of its capabilities - taking a given
graph (in our case, the galaxy.stars graph) and returning
_all_ the connected components existing in said graph?

After converting each component in the resulting list to
a set, we can then take advantage of Python's O(1)
set-membership checks to efficiently filter out the
doomed-to-fail pathfinding checks that start and end in
different graph components.

For this attempt, I've used the largest few components
(usually largest 3-6) to strike a good balance between
doomed-route coverage and effort.