Add initial cut at delta debugging
As PyRoute consumes sector files, I've specialised this version to
take advantage of the structure of those files, to save travelling
down some blind alleys.  For the moment, the moment, delta
debugging can process and reduce only a single sector file.

This includes a reset-button method to take a Galaxy
object back to a bare state suitable for the next iteration of delta
debugging, rather than pay the cost of starting afresh each time.

Additionally, I've modified write_routes to take an optional item
delimiter (instead of the default space), to make reading routes
back in easier.

For example, minimising a Dagudashaag sector run that tripped the
assertion added in the last PR (#27) started with 559 stars (as per
TravellerMap download, c 06 Dec 2021), and finished with 16.