Verify 1-minimality in single-file delta debugging
An interesting input is k-minimal if no subset of that input,
of at most k items, can be removed without making it
uninteresting.

As the number of runs needed to experimentally verify k-minimality
varies as O(N ** k) for input size N, settling for k=1 seems to give
a good compromise between level of input shrinkage and run time needed.