Add delta debugging-specific Galaxy variant.
Since the DeltaDictionary contains the current reduction state,
I needed a way to turn that into what the core PyRoute gubbins
expects.

I found the simplest way to do that was to add a Galaxy variant
that sawed off the original's read_sectors() method and added
its own version that unpacked a DeltaDictionary.