Add specific-line removal tu support line reduction
Unlike (sub)sector reduction, where the supporting
primitive was _taking_ a subset of the existing input,
the easiest way I could scheme up to support line
reduction was by _removing_ a subset of the existing
input.

There's probably a quicker and more pythonic way
to do this, but I didn't want the make perfect be
the enemy of good.