Add line-level reduction
Using the primitive added in previous commits, line-level reduction
treats individual star lines in the sector file as the unit of
reduction.  This allows more fine-grained reductions than (sub)sector
reduction, at the cost of more pathfinding calls on the same input.

As such, line reduction is intended to be called _after_ the coarser
reductions.  Setting the singleton_only flag to True lets the
caller attain and verify that their input is 1-minimal - each star
line of that input is individually necessary to reproduce the
issue being pursued.

The Dagudashaag-subsecteor-spiked.sec file in Tests/DeltaFiles is
the result of running subsector reduction on the original
Dagudashaag-spiked.sec file, to show an example of how reduction
should work and to provide a smaller starting point for the
line reduction tests.