Add is_uninteresting check to reducer.
This is the logical dual of DeltaReduce.is_initial_state_interesting().

It retains the interestingline/interestingtype sensitivity, and is
intended to be used in regression tests to verify no exception of
interest (as specified in interestingline/interestingtype) has been
thrown.

When an exception is specified, this can be used in a regression test
that verifies that the underlying problem that caused that exception
_is_ fixed and _stays_ fixed.

With no exception specified, is_initial_state_uninteresting() serves
as a general "nothing that we can detect went boom" check.