Hook up TradeBalance in lieu of dict
As the machinery for trade balancing was duplicating that for
pax balancing, I'm bowing to the inevitable and pulling that
out into a specialist object that can be seperately tested.

As TradeBalance extends dict, the first step is to check that
it's a drop-in replacement, and fix the resulting circular
import.