Add intra-star reduction machinery
Reaching a 1-minimal set of star lines is massive progress from
whatever the user started with, but may not be sufficient for them
to isolate their problem.

There's still a fair amount of possibly-irrelevant stuff inside a
given starline.  Rather than blithely assume that's all relevant,
_test_ that assumption by trying some or all of the following:
  Removing {trade,noble,base} codes
  Removing amber/red zone flag
  Removing base codes
  Removing extra stars
  Removing route notes (eg Xboat routes)
  Setting system TL to 8
  Setting System port to C
  Setting PBG to 100
  Removing extra worlds

The goal is to further ease debugging by reducing the remaining
amount of irrelevant stuff the user has to wade through and get
their head around.