Add switch_lines primitive to DeltaDictionary
The caller needs to supply a list of 2-element tuples - first in each
tuple is the original line to be replaced, and the second element is
the line to replace the corresponding original _with_.  Switch_lines
assumes each tuple element will result in a well-formed Star object.

This is intended to support within-starline reduction.