Delta debug remix (#34)
* Add initial debugging-support classes

* Disambiguate intramodule imports

* Add load_from_traveller_map method to SectorDictionary

This commit follows the lead of the Star object's
static parse_line_into_star method to provide an easy
way to turn an on-disk sector file into an in-memory
SectorDictionary, ready for use in delta debugging.

* Cache regex lookups when parsing lines to Stars

* Track allegiances during file load

* Add subset generation to DeltaDictionary

Ultimately, I intend delta debugging to support three levels of
reduction:
- by sector
- by subsector
- by individual star

To keep concerns cleanly separated, the DeltaDictionary will
handle the bookkeeping of the input being reduced, while delta
debugging proper will track the reduction itself.

This commit adds the support for the first two levels.

* Add delta debugging-specific Galaxy variant.

Since the DeltaDictionary contains the current reduction state,
I needed a way to turn that into what the core PyRoute gubbins
expects.

I found the simplest way to do that was to add a Galaxy variant
that sawed off the original's read_sectors() method and added
its own version that unpacked a DeltaDictionary.

* Update references

* Verify sector position in subset methods

* Fix subsector_subsect to match Sector class assumptions

As the existing code (specifically, Sector.set_bounding_subsectors)
implicitly assumes all sixteen subsectors exist in a subsector,
this clahses with the earlier assumption made by subsector_subset
that unneeded subsectors can be omitted.

Since this broke read-from-subsectors, subsector_list now returns
omitted subsectors as empty.

* Add initial reduction engine

The reduction engine actually makes the decisions about how to
generate candidate reductions to see which one(s) are interesting.

For this initial commit, I added the core interestingness test
(does the supplied input still trip the desired bug?), an initial
check (no point continuing if your initial input doesn't trip
the bug you're looking for), and a subsector-level reduction pass.

As the name suggests, the subsector reduction pass attempts to
reduce the input by treating complete subsectors as the unit of
reduction.

My overall intention for this chunk of work is to
add sector-level and line-level reduction alongside subsector
reduction, and call them in descending unit size - sector first,
then subsector, and finally line.  The earlier passes then remove
larger chunks of irrelevant input, speeding up the later passes.

* Add specific-line removal tu support line reduction

Unlike (sub)sector reduction, where the supporting
primitive was _taking_ a subset of the existing input,
the easiest way I could scheme up to support line
reduction was by _removing_ a subset of the existing
input.

There's probably a quicker and more pythonic way
to do this, but I didn't want the make perfect be
the enemy of good.

* Add SectorDictionary fields missed during subsector listing

* Add line-level reduction

Using the primitive added in previous commits, line-level reduction
treats individual star lines in the sector file as the unit of
reduction.  This allows more fine-grained reductions than (sub)sector
reduction, at the cost of more pathfinding calls on the same input.

As such, line reduction is intended to be called _after_ the coarser
reductions.  Setting the singleton_only flag to True lets the
caller attain and verify that their input is 1-minimal - each star
line of that input is individually necessary to reproduce the
issue being pursued.

The Dagudashaag-subsecteor-spiked.sec file in Tests/DeltaFiles is
the result of running subsector reduction on the original
Dagudashaag-spiked.sec file, to show an example of how reduction
should work and to provide a smaller starting point for the
line reduction tests.

* Flag excluded subsectors separately.

The previous approach, counting empty subsectors as excluded,
clashed with actual data (such as Reft sector) that contained
void (ie, empty) subsectors.  As such, this commit changes to
marking excluded subsectors by setting their items field to
None, instead of an empty list.

This caused some minor blast damage, also fixed in this commit.

* Add sector-level reduction

Finish out the initial three reduction passes by adding sector-level
reduction.  As sectors are the coarest reduction units used thus far,
it is intended to be called first.

* Make raw_lines assembly more pythonic

* Dedupe raw_lines assembly

* Exclude skipped subsectors from line-drop processing

* Add skipped property to SectorDictionary

* Dedupe args setup in reducer tests

* Verify line reduction can drop empty sectors

* Separate out singleton-line test

* Fix singleton detection in line reduction

* Add 2-minimality line-reduction pass

There is (of course) no guarantee that a 1-minimal input set is
small enough to be tractable to manual analysis.

The next step beyond 1-minimality is, strangely enough,
2-minimality - ensuring that removing _any_ set of 2 input elements
breaks interestingness.

As enforcing 2-minimality requires O(N**2) pathfinding runs, it
is intended (and strongly recommanded) to invoke 2-minimisation on
an already 1-minimal input.

By excluding (0,1), (1,2) ... (0,2), (1,3) .. pairs of input
elements rather than (0,1), (0,2) ... (1,2), (1,3) ... , I'm
trying to avoid getting hung up on an early input element that
won't get removed.

* Fix issues found by delta debugging

* Add delta debugging driver class

* Assert equality of forward and reverse route costs

* Deep six obsolete TODO

* Assert equality of forward and reverse route costs

* Display end-of-pass shortest error message

* Add well-formed check to Galaxy and Star classes

* Add Star deep copy

* Add config options to skip (sub)sector reduction

* Display (sub)sector counts when reducing

* Add is_uninteresting check to reducer.

This is the logical dual of DeltaReduce.is_initial_state_interesting().

It retains the interestingline/interestingtype sensitivity, and is
intended to be used in regression tests to verify no exception of
interest (as specified in interestingline/interestingtype) has been
thrown.

When an exception is specified, this can be used in a regression test
that verifies that the underlying problem that caused that exception
_is_ fixed and _stays_ fixed.

With no exception specified, is_initial_state_uninteresting() serves
as a general "nothing that we can detect went boom" check.

* Add run-line option to driver class

* Simplify post-run followup

* Add option to assume starting input is interesting

* Zero allegiance stats when cloning SectorDictionary objects

As the various methods that clone SectorDictionary are used
in setting up a _new_ pathfinding run, not clearing the
previously-accumulated passenger and trade totals is a problem.

That problem is now fixed in this commit.

* Sort DeltaDictionary.subsector_list output

* Drop superfluous pass line

* Fix test blast damage

* More review-rumbled cleanup

* Move delta debug bits into their own subfolder

* Rumble route_cost b0rkage

* Update running total in route_cost

* Add debug flag to Gaalxy, TradeCalc

* Add cli option to set debug flag

* Characterise sector-subset blowup on empty Vland

* Add Allegiance.__deepcopy__ method

* Characterise objectStatistics blowup

* Add ObjectStatistics.__deepcopy__ method

* Fix local import

* Check removed item is regenerated