De-duplicate queued items on new upper bound
We're already discarding queued items that bust the new upper bound,
and new upper bounds are found fairly infrequently.

We're now taking the opportunity to remove some of the queued items
that are doomed to run afoul of the "qcost <= ncost" check in the
if-neighbour-enqueued block without having to grind through them
one at a time.  Trimming the queue further also speeds up operations
whose runtime is a function of N, such as heapifying, pushing, popping,
etc.