Add Hex class, prep for Star simplification
As at this commit, the Star class convolves astrographic position,
economic data, and other things in one place.  That may have made
sense initially, when the Star class was much smaller than it is
today, but not any longer.

The Hex class aims to encapsulate all the astrographic location/
distance concerns away from the rest of the Star class.