Add multiproc version of TradeCalculation
This adds a new performance enhancement version of the TradeCalculation.
Using the Python Multiprocessor module to split the route finding, the expensive part of the program run, into multiple processes.
This enables a 50% or better improvement in the overall time to process a set of stars and routes.
There are some limitations, like with any optimisation, trading (in this case) more memory space for less time.

Adding this needs two changes to the route.py: to add an option to one command line parameter, and one new command line option.
Then modify the route -> Galaxy API to move several configuration parameters to the lowest level possible to optimise their use.
These changes then ripple through the Delta Debugging and test classes.

One attempt to run the tests uncovered a broken import, so fix that.

Update the documentation to discuss using the MP version of trade generator.

- Fix the now broken tests