Add widen-hole reduction pass
The "hole" implied in the reducer name is the one caused by dropping
out 1 or more sectors, subsectors, or lines.  While the new hole
information is available, widen the breach by iteratively trying to
remove 1, 2, 4, 8, ... lines from one side of the breach, then the
other.

If there's not much detritus adjacent to the new hole, widening will
bail out quickly and the overall reduction process trundles onwards.
If there _is_ a lot of detritus, the iterated doubling mentioned
above can clean up a lot of it, potentially speeding up the
overall reduction.