Add more delta debug passes (#64)
* Characterise ehex_to_int and add reverse mapping

* Add Star.parse_to_line

* Add more fields to Star constructor

* Add intra-star reduction machinery

Reaching a 1-minimal set of star lines is massive progress from
whatever the user started with, but may not be sufficient for them
to isolate their problem.

There's still a fair amount of possibly-irrelevant stuff inside a
given starline.  Rather than blithely assume that's all relevant,
_test_ that assumption by trying some or all of the following:
  Removing {trade,noble,base} codes
  Removing amber/red zone flag
  Removing base codes
  Removing extra stars
  Removing route notes (eg Xboat routes)
  Setting system TL to 8
  Setting System port to C
  Setting PBG to 100
  Removing extra worlds

The goal is to further ease debugging by reducing the remaining
amount of irrelevant stuff the user has to wade through and get
their head around.

* Add switch_lines primitive to DeltaDictionary

The caller needs to supply a list of 2-element tuples - first in each
tuple is the original line to be replaced, and the second element is
the line to replace the corresponding original _with_.  Switch_lines
assumes each tuple element will result in a well-formed Star object.

This is intended to support within-starline reduction.

* Dedupe new-sector-dictionary setup

* Add protoype new-style reduction pass, Canonicalisation

After some more skull sweat, I realised there was an
easier way to add reduction passes.  The DeltaReduce
object ultimately needs to be hollowed out to
drive a supplied list of reduction passes.

Each pass instance requires a DeltaReduce object,
to provide support gubbins common to reduction passes.

The idea of the preflight() method is to check that
whetever needed preconditions hold before investing
time in attempting the reduction pass.  The run()
method actually runs the reduction pass, and
updates the reducer's dictionary with whatever
reduction got found.

* Add full-reduction pass

* Fix zone and base clearing

* Consolidate reduce-all tests

* Trim lines when reducing

* Wring out full-reduction pass

* Add auxiliary-line input reducer

* Deduplicate within-line reduction passes

* Hoist out subs_list generation

* Harmonise status messages

* Deduplicate run methods

* Fix circular import

* Hook within-line reduction to delta debug

* Fix issues rumbled by integration testing

* Isolate and fix non-auxiliary-reduction

* Test sophont-list edge case

* Add sophont-in-trade-code reduction to DeltaStar

* Fix aux-reduction subs-list generation

* Hook up sophont reduction

* Fix world-count reduction

* Add capital-reduction pass

This is broken out seperately because of the quite disparate effects
capital-reduction will have on trade and comm/xboat maps.

* Add mp-threads parameter to DeltaDebug

* Hoist out sector reduction

* Hoist out subsector reduction

* Hoist out single-line reduction

* Hoist out two-line reduction

* Add widen-hole reduction pass

The "hole" implied in the reducer name is the one caused by dropping
out 1 or more sectors, subsectors, or lines.  While the new hole
information is available, widen the breach by iteratively trying to
remove 1, 2, 4, 8, ... lines from one side of the breach, then the
other.

If there's not much detritus adjacent to the new hole, widening will
bail out quickly and the overall reduction process trundles onwards.
If there _is_ a lot of detritus, the iterated doubling mentioned
above can clean up a lot of it, potentially speeding up the
overall reduction.

* Hook up reverse breach widening to line reduction

* Add allegiance_list primitive

* Add allegiance_subset primitive

* Add allegiance-reduction pass

* Add allegiance-reduction argument

* Squash ruff complaints

* Fix CI-breaking alignment goof

* Unshadow outer-scope gubbins in map-remap

* Fix test import

* Fix widen-hole reversals

* Length-guard line chunk removals

* Characterise trade-code reduction

* Characterse route-link reduction

* Characterise within-line reduction

* Fix test blast damage

* Update sector allegiances in headers after reduction

* Drop empty-allegiance-list sectors when taking allegiance subsets

* Bail out of within-line reductions early if nothing to do

* Add reduce-full-line option to reducer

* Square up hex block when canonicalising lines

* Add progress indicators to main reduction passes

* Add end-of-lines reduction option