Property-ise UWP fields
This resolves the multiple-point-of-truth problem that those chunks
previously had - it was possible for the ehex and int representations
to diverge.  Now, both versions (of, eg, size) are generated off the
same underlying field, obviating that risk.

Excepting the port UWP value, FOO delivers the _ehex_ version of the value,
and FOO_CODE delivers the integer version.