Add indirect option to min_cost retrieval
Min-cost vector originally gave a way to exclude a node's neighbours
during pathfinding that were too close - less than the minimum-edge
cost for that node - from a finite upper bound.

For smaller route-reuse values, this has worked like a charm.  Not
so well for larger values, as I've had to pull back the number of
landmarks per component.

The next obvious step up from "closest immediate neighbour" to node
is "closest non-immediate neighbour" to node - ie, minimum of sum
of the two involved edges.  That's quadratic in the number of
immediate neighbours, so I've gone with sum of minima as a lower
bound for that minimum of sum.