Filter trade's raw_routes by BTN upper bound
The get_btn calculation is somewhat heavyweight, being (roughly)

star1's WTN + star2's WTN + (1 if one is ag, other needs ag) +
(1 if one is in, other needs in) + (-1 if not allies) +
(-1 if either or both is Wild) + (offset for distance),
then capped by maximal BTN given star WTNs.

A comparatively simple upper bound on that is assuming that
both (one is ag & other needs ag), and (one is in & other needs in),
both systems are allied, and skipping the maximal-BTN cap.

Modulo systems within max-jump range of each other, if that upper
bound on BTN doesn't meet/beat the minimum BTN to be considered,
there's no point calculating the _actual_ BTN, as we now know it
won't meet/beat the minimum BTN, either.