Trim out surplus raw ranges
As itertools.combinations(foo, 2) scales as N(foo)**2, we can
take advantage of that in two ways.
1.  First, remove red-zoned stars from self.galaxy.ranges.
2.  Second, split the remaining stars into a hiball set, where their
WTN meets/beats the minimum WTN, and a loball set, where their WTN's
don't.

The original full comprehension now only runs on the combinations
over the hiball set.  The loball combinations, and the hiball-loball
product, where at least one element of the pair has WTN below
the minimum WTN, only need to be checked for the pair to be within
maximum jump range.

In testing, the result was a roughly ~25% saving in jump generation
time.