Tighten BTN upper bound by capping it to max_btn
Pulling in the get_max_btn call, similarly as for get_btn,
tightens the BTN upper bound where the two WTNs are mismatched
enough to make the smaller system WTN the binding constraint on
what the (actual or upper bound) BTN _can_ be.

For some reason, this results in a ~16% speedup, compared to
status quo, on trade jump generation for the full-orchestra.