Hook up allegiance count verification to galaxy
When a star is inserted, it's added to the allegiances for galaxy,
sector and subsector.  As such, the galaxy-level allegiance
world counts should should match the sum of the sector-level allegiance
world counts, otherwise the galaxy is not well-formed.

Because integer equality is transitive (if A == B, and B == C, then
A == C), the previous commit (sector-level allegiance count
verification) combined with this one means the sum of subsector
allegiance counts doesn't have to be compared directly with
galaxy level counts.