Zero-guard large_btn_index
It's possible, if unlikely, that the largest BTN in the list of
candidate routes is below 18.  Previously, this would cause a
StopIteration error.

To fix that error, first initialise large_btn_index to 0, as a
safe default, then IF and ONLY IF the first route's BTN is 18+,
call the original code.