Defang uploadWorlds in WikiUploadPDF
Per discussion with @tjoneslo, "The basic idea is sound.  The list of 5
sectors used for the update is on line 310.  The plan was to eventually
update all of the world pages.  But enough other people have gone
through and set up the world pages the page template formatting is
going to be destructive rather than helpful.  And without the page
template, the data data template doesn't have much use."

In light of the above, defang uploadWorlds to turn it into a no-op.